requirements.txt
astroid==2.6.6
attrs==21.2.0
iniconfig==1.1.1
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
numpy==1.21.2
packaging==21.0
pluggy==0.13.1
py==1.10.0
pylint==2.9.6
pyparsing==2.4.7
pytest==6.1.2
PyYAML==5.4.1
toml==0.10.2
wrapt==1.12.1
